avoid memleak exceed int range when using heap-checker
[alkondratenko@gmail.com] format size_t with %zu

+++ src/heap-profile-table.cc
@@ -517,7 +517,7 @@ HeapProfileTable::Snapshot* HeapProfileTable::NonLiveSnapshot(
 // Information kept per unique bucket seen
 struct HeapProfileTable::Snapshot::Entry {
   int count;
-  int bytes;
+  size_t bytes;
   Bucket* bucket;
   Entry() : count(0), bytes(0) { }
 
@@ -592,7 +592,7 @@ void HeapProfileTable::Snapshot::ReportLeaks(const char* checker_name,
   for (int i = 0; i < to_report; i++) {
     const Entry& e = entries[i];
     base::RawPrinter printer(buffer, kBufSize);
-    printer.Printf("Leak of %d bytes in %d objects allocated from:\n",
+    printer.Printf("Leak of %zu bytes in %d objects allocated from:\n",
                    e.bytes, e.count);
     for (int j = 0; j < e.bucket->depth; j++) {
       const void* pc = e.bucket->stack[j];
